+++ composer.json
@@ -36,7 +36,7 @@
     },
     "require": {
         "php": "^7.2 | ^8.0",
-        "vimeo/psalm": "^4.7.3"
+        "vimeo/psalm": "^4.8"
     },
     "prefer-stable": true,
     "archive": {
